\name{haspelmath}
\alias{haspelmath}
\docType{data}
\title{
Data about indefinite constructions in 39 different languages
}
\description{
Summary of the data in Haspelmath (1997), classical example of the usage of semantic maps.
}
\usage{data(haspelmath)}
\format{
The dataset is a matrix with nine rows, describing the different indefinite functions, and 134 columns documenting for each individual construction which functions are possibly expressed by it.
}
\source{
Haspelmath, Martin. Indefinite Pronouns. Oxford Studies in Typology and Linguistic Theory. Oxford: Clarendon, 1997.
}
\examples{
data(haspelmath)
## English data
haspelmath[,7:9]
}
\keyword{datasets}
