\name{qlcVisualize-package}
\alias{qlcVisualize-package}
\alias{qlcVisualize}
\docType{package}
\title{
Visualizations for Quantitative Language Comparison
}
\description{
A collection of specific visualisations of data as used in quantitative language comparison.
}
\details{
\tabular{ll}{
Package: \tab qlcVisualize\cr
Type: \tab Package\cr
Version: \tab 0.4\cr
Date: \tab 2024-07-31\cr
License: \tab GPL-3\cr
}
Currently implemented are visualisations dealing nominal data with multiple levels \link{lmap} ("levelMap") and \link{fmap} ("factorMap"), and some assistance for making of weighted geographic Voronoi maps \link{wmap} ("weigthedMap").
}
\author{
Michael Cysouw <cysouw@mac.com>
}
