\name{qlcVisualize-package}
\alias{qlcVisualize-package}
\alias{qlcVisualize}
\docType{package}
\title{
Visualizations for Quantitative Language Comparison
}
\description{
A collection of specific visualisations of data as used in quantitative language compariseon.
}
\details{
\tabular{ll}{
Package: \tab qlcVisualize\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2024-05-08\cr
License: \tab GPL-3\cr
}
Currently implemented are visualisations dealing nominal data with multiple levels \link{lmap} ("level-map") and \link{limage} ("level-image"), and some assistance for making geographic Voronoi maps \link{vmap}.
}
\author{
Michael Cysouw <cysouw@mac.com>
}
